requirements.txt
aenum==3.1.5
altgraph==0.17.2
asttokens==2.0.5
autopep8==1.6.0
backcall==0.2.0
backports.entry-points-selectable==1.1.1
certifi==2021.10.8
colorama==0.4.4
distlib==0.3.3
Django==3.2
django-filter==21.1
djangorestframework==3.13.1
et-xmlfile==1.1.0
executing==0.8.3
filelock==3.4.0
flake8==4.0.1
future==0.18.2
httplib2==0.20.4
idna==3.3
ipython==8.1.1
jedi==0.17.2
jws==0.1.3
matplotlib-inline==0.1.3
mccabe==0.6.1
mypy==0.931
mypy-extensions==0.4.3
numpy==1.21.4
openpyxl==3.0.9
parso==0.7.1
pefile==2022.5.30
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.0
pluggy==1.0.0
prompt-toolkit==3.0.28
protobuf==3.20.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycryptodome==3.14.1
pyflakes==2.4.0
Pygments==2.11.2
pyinstaller==5.3
pyinstaller-hooks-contrib==2022.8
pyls==0.1.6
pyparsing==3.0.7
Pyrebase4==4.5.0
pytesseract==0.3.8
python-jsonrpc-server==0.4.0
python-jwt==2.0.1
python-language-server==0.36.2
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.3.3
pytz==2021.3
pywin32-ctypes==0.2.0
repath==0.9.0
requests==2.27.1
requests-toolbelt==0.9.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
toml==0.10.2
tomli==2.0.0
traitlets==5.1.1
typing_extensions==4.0.1
tzdata==2021.5
ujson==5.1.0
urllib3==1.26.9
virtualenv==20.10.0
virtualenvwrapper-win==1.2.6
wcwidth==0.2.5
websocket-client==1.3.3
